Implementada la funcionalidad de la parte 2.

+++ src/Aritmetica_Logica/Fraccion.java
@@ -5,14 +5,14 @@ package Aritmetica_Logica;
  * el 10-Mar-17.
  */
 public class Fraccion {
-    public int numerador;
-    public int denominador;
+    public long numerador;
+    public long denominador;
 
-    public Fraccion(int _numerador, int _denominador){
+    public Fraccion(long _numerador, long _denominador){
         numerador = _numerador;
         denominador = _denominador;
     }
-    public Fraccion(int _numerador){
+    public Fraccion(long _numerador){
         numerador = _numerador;
         denominador = 1;
     }
@@ -22,8 +22,8 @@ public class Fraccion {
     }
     //Operaciones basicas
     public static Fraccion sumar(Fraccion operando1, Fraccion operando2){
-        int numeradorResultado;
-        int denominadorResultado = econtrarDenominadorComun(operando1.denominador, operando2.denominador);
+        long numeradorResultado;
+        long denominadorResultado = econtrarDenominadorComun(operando1.denominador, operando2.denominador);
 
         numeradorResultado = (operando1.numerador * (denominadorResultado/operando2.denominador)) +
                 (operando2.numerador * (denominadorResultado / operando2.denominador));
@@ -32,8 +32,8 @@ public class Fraccion {
     }
 
     public static Fraccion restar(Fraccion operando1, Fraccion operando2){
-        int numeradorResultado;
-        int denominadorResultado = econtrarDenominadorComun(operando1.denominador, operando2.denominador);
+        long numeradorResultado;
+        long denominadorResultado = econtrarDenominadorComun(operando1.denominador, operando2.denominador);
 
         numeradorResultado = (operando1.numerador * (denominadorResultado/operando2.denominador)) -
                 (operando2.numerador * (denominadorResultado / operando2.denominador));
@@ -42,42 +42,42 @@ public class Fraccion {
     }
 
     public static Fraccion multiplicar(Fraccion operando1, Fraccion operando2){
-        int numeradorResultado = operando1.numerador * operando2.numerador;
-        int denominadorResultado = operando1.denominador * operando2.denominador;
+        long numeradorResultado = operando1.numerador * operando2.numerador;
+        long denominadorResultado = operando1.denominador * operando2.denominador;
 
         return simplificar(new Fraccion(numeradorResultado,denominadorResultado));
     }
 
     public static Fraccion multiplicar(Fraccion operando1, int operando2){
-        int numeradorResultado = operando1.numerador * operando2;
-        int denominadorResultado = operando1.denominador * 1;
+        long numeradorResultado = operando1.numerador * operando2;
+        long denominadorResultado = operando1.denominador * 1;
 
         return simplificar(new Fraccion(numeradorResultado,denominadorResultado));
     }
 
 
     public static Fraccion dividir(Fraccion operando1, Fraccion operando2){
-        int numeradorResultado = operando1.numerador * operando2.denominador;
-        int denominadorResultado = operando1.denominador * operando2.numerador;
+        long numeradorResultado = operando1.numerador * operando2.denominador;
+        long denominadorResultado = operando1.denominador * operando2.numerador;
 
         return simplificar(new Fraccion(numeradorResultado, denominadorResultado));
     }
     public static Fraccion dividir(int operando1, Fraccion operando2){
-        int numeradorResultado = operando1 * operando2.denominador;
-        int denominadorResultado = 1 * operando2.numerador;
+        long numeradorResultado = operando1 * operando2.denominador;
+        long denominadorResultado = 1 * operando2.numerador;
 
         return simplificar(new Fraccion(numeradorResultado, denominadorResultado));
     }
     public static Fraccion simplificar(Fraccion fraccion){
-        int divisorComun = encontrarDivisorComun(fraccion.numerador, fraccion.denominador);
-        int numeradorResultado = fraccion.numerador / divisorComun;
-        int denominadorResultado = fraccion.denominador / divisorComun;
+        long divisorComun = encontrarDivisorComun(fraccion.numerador, fraccion.denominador);
+        long numeradorResultado = fraccion.numerador / divisorComun;
+        long denominadorResultado = fraccion.denominador / divisorComun;
 
         return new Fraccion(numeradorResultado, denominadorResultado);
     }
 
     //Auxiliares
-    private static int econtrarDenominadorComun(int denominador1, int denominador2){
+    private static long econtrarDenominadorComun(long denominador1, long denominador2){
         if(denominador1 == denominador2){
             return denominador1;
         } else {
@@ -94,9 +94,9 @@ public class Fraccion {
     }
 
     //Encuentra el maximo divisor comun entre dos numeros, se usa para simplificar fracciones.
-    private static int encontrarDivisorComun(int numero1, int numero2){
+    private static long encontrarDivisorComun(long numero1, long numero2){
         int divisor = 2;
-        int minimo = numero1;
+        long minimo = numero1;
         if (numero1 > numero2)
             minimo = numero2;
         while (!(divisor > minimo)){
